* commit '72f99531b7cdfad1c3fca02d32e88f4ac2662043':
  Core/Players: Update the location of the cinematic camera only if there is an active cinematic (#26215)
  Fix build

+++ src/server/game/Handlers/AdventureMapHandler.cpp
@@ -29,7 +29,7 @@ void WorldSession::HandleAdventureMapStartQuest(WorldPackets::AdventureMap::Adve
 
     auto itr = std::find_if(sAdventureMapPOIStore.begin(), sAdventureMapPOIStore.end(), [&](AdventureMapPOIEntry const* adventureMap)
     {
-        return adventureMap->QuestID == startQuest.QuestID && _player->MeetPlayerCondition(adventureMap->PlayerConditionID);
+        return adventureMap->QuestID == uint32(startQuest.QuestID) && _player->MeetPlayerCondition(adventureMap->PlayerConditionID);
     });
 
     if (itr == sAdventureMapPOIStore.end())
